python/data/requirements.txt
grpcio==1.10.0
pika==0.11.2
protobuf==3.5.2
pymongo==3.6.1
six==1.11.0
